feat: add jwt_vc_json validationto batch credentials
